refactor(compiler): split emit-init-sections.ts into one file per phase (B3)
B3 of the post-#1054 emit-init maintainability plan. Following B1's
declarative `PHASES` registry (#1058), break the 236-line
`emit-init-sections.ts` "miscellaneous bucket" into 10 single-purpose
files under `phases/` — one per emit phase, plus one for the shared
pure helper.

Changes:

- New `phases/conditional-slot-ids.ts` — pure helper
  `collectConditionalSlotIds`, kept separate from the side-effecting
  emit phases.
- New `phases/{props-extraction,props-event-handlers,rest-attr-applications,
  event-handlers,ref-callbacks,effects-and-on-mounts,init-statements,
  provider-and-child-inits,static-array-child-inits}.ts` — each carries
  exactly one `emit*` function with its own focused header comment.
- The trailing-blank-line behaviour previously open-coded in the
  orchestrator (`if (ctx.initStatements.length > 0) lines.push('')`)
  is folded into `phases/init-statements.ts`. The phase is now
  self-contained; `phases.ts`'s entry shrinks to a one-liner.
- `phases.ts` imports each phase from its own file; `element-refs.ts`
  picks up `collectConditionalSlotIds` from the new helper file.
- Deleted `emit-init-sections.ts`. Updated three doc-comment references
  (`control-flow.ts`, `plan/declaration-emit.ts`, `stringify/declaration-emit.ts`).

Each phase file is now 22–70 lines with a single concern. Adding a new
phase requires creating one new file + appending one entry to `PHASES`
— no risk of accidentally cross-mixing pure / side-effecting helpers
in a shared bucket.

Output stays byte-identical: jsx (783 / 4 unrelated resolver-alias
fails unchanged) + adapter + client (426 / 0) tests all green.
